requirements.txt
biopython>=1.84
datasets>=2.18
einops>=0.8
huggingface_hub>=0.30
numpy>=1.26
qdrant-client>=1.11
torch>=2.6
transformers>=4.48
